Move off EllipsizeParams as a static utility class ... as it starts to take up a great deal of space
